ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
scikit-learn==1.8.0
joblib==1.5.3
numpy==2.4.3
mlflow==2.22.0
sqlalchemy>=2.1.0
python-jose[cryptography]==3.3.0

python-multipart==0.0.6
email-validator==2.1.0

bcrypt>=4.0
